packages/harpguru-core/src/components/harp-cell/hooks/use-add-buffered-activity-toggle/use-add-buffered-activity-toggle.ts
import { useGlobal } from 'reactn'
import type { DegreeIds } from 'harpparts'

type BufferToggleFunction = (arg0: DegreeIds) => void

export const useAddBufferedActivityToggle = (): BufferToggleFunction => {
  const [bufferedActivityToggles, setBufferedActivityToggles] = useGlobal(
    'bufferedActivityToggles'
  )

  return (toggleToBuffer) => {
    if (bufferedActivityToggles.includes(toggleToBuffer)) {
      setBufferedActivityToggles([...bufferedActivityToggles])
    } else {
      setBufferedActivityToggles([...bufferedActivityToggles, toggleToBuffer])
    }
  }
}
